webrtc-stats: add track handler with tests
