+++ spec/unit/webrtc/stats/mediaSsrcHandler.spec.ts
@@ -13,28 +13,28 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-import { Mid, Ssrc, TransceiverHandler } from "../../../../src/webrtc/stats/transceiverHandler";
-import { description } from "./sdp";
+import { Mid, Ssrc, MediaSsrcHandler } from "../../../../src/webrtc/stats/mediaSsrcHandler";
+import { REMOTE_SFU_DESCRIPTION } from "../../../test-utils/webrtc";
 
-describe("TransceiverHandler", () => {
+describe("MediaSsrcHandler", () => {
     const remoteMap = new Map<Mid, Ssrc[]>([
         ["0", ["2963372119"]],
         ["1", ["1212931603"]],
     ]);
-    let handler: TransceiverHandler;
+    let handler: MediaSsrcHandler;
     beforeEach(() => {
-        handler = new TransceiverHandler();
+        handler = new MediaSsrcHandler();
     });
     describe("should parse description", () => {
         it("and build mid ssrc map", () => {
-            handler.parse(description, "remote");
+            handler.parse(REMOTE_SFU_DESCRIPTION, "remote");
             expect(handler.getSsrcToMidMap("remote")).toEqual(remoteMap);
         });
     });
 
     describe("should on find mid by ssrc", () => {
         it("and return mid if mapping exists.", () => {
-            handler.parse(description, "remote");
+            handler.parse(REMOTE_SFU_DESCRIPTION, "remote");
             expect(handler.findMidBySsrc("2963372119", "remote")).toEqual("0");
         });
     });
